Update to Qt 6.8 base (#16)
* transformenums.go: Add type resolution for enums and flags

* transformstructs.go: Add type resolution for inner classes and structs

* clangexec.go: Remove dead code

* transformtypedefs.go: Expand type resolution

* statetracker.go: Expand type registration and remove brute force types

* clang2il.go: Expand type resolution for methods

* config-allowlist.go: Update quirks and allowed types and headers

* intermediate.go: Update helper methods

* config-libraries.go, main.go: Update declaration and invocation for modules

* emitcabi.go: Add support for Qt 6.8 base and remove dead code

* emitzig.go: Add support for Qt 6.8 base

* Update bindings and wrappers

* go.mod: Update Go module properties

* build.zig.zon: Update to Qt 6.8.2

* README: Update for Qt 6.8

+++ include/libqbuffer.h
@@ -38,7 +38,7 @@ libqt_string QBuffer_Tr(const char* s);
 libqt_string QBuffer_Buffer(QBuffer* self);
 libqt_string QBuffer_Buffer2(const QBuffer* self);
 void QBuffer_SetData(QBuffer* self, const libqt_string data);
-void QBuffer_SetData2(QBuffer* self, const char* data, int lenVal);
+void QBuffer_SetData2(QBuffer* self, const char* data, ptrdiff_t lenVal);
 libqt_string QBuffer_Data(const QBuffer* self);
 bool QBuffer_Open(QBuffer* self, int openMode);
 void QBuffer_OnOpen(QBuffer* self, intptr_t slot);
